FIX major issues for inventory - Move all needed declaration from coreobject to commonobject for updateCommon, fetchCommon, deleteCommon, createCommon - Just leave method for child object into coreobject, check later if is useless - debug inventory card, PHP 7.0 compatibility - activate rights for inventory into stock module - add inventory options into stock admin module
