+++ htdocs/core/class/commonobject.class.php
@@ -11,6 +11,7 @@
  * Copyright (C) 2012      Cedric Salvador      <csalvador@gpcsolutions.fr>
  * Copyright (C) 2015      Alexandre Spangaro   <aspangaro.dolibarr@gmail.com>
  * Copyright (C) 2016      Bahfir abbes         <dolipar@dolipar.org>
+ * Copyright (C) 2017      ATM Consulting       <support@atm-consulting.fr>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -4662,7 +4663,159 @@ abstract class CommonObject
 		return $buyPrice;
 	}
 	
+	/**
+	 * Function test if type is date
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_date($info)
+	{
+		if(isset($info['type']) && $info['type']=='date') return true;
+		else return false;
+	}
 	
+	/**
+	 * Function test if type is array
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_array($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['type']) && $info['type']=='array') return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function test if type is null
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_null($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['type']) && $info['type']=='null') return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function test if type is integer
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_int($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['type']) && ($info['type']=='int' || $info['type']=='integer' )) return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function test if type is float
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_float($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['type']) && $info['type']=='float') return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function test if type is text
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_text($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['type']) && $info['type']=='text') return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function test if is indexed
+	 *
+	 * @param   array   $info   content informations of field
+	 * @return                  bool
+	 */
+	protected function is_index($info)
+	{
+		if(is_array($info))
+		{
+			if(isset($info['index']) && $info['index']==true) return true;
+			else return false;
+		}
+		else return false;
+	}
+	
+	/**
+	 * Function to prepare the values to insert
+	 *
+	 * @return array
+	 */
+	private function set_save_query()
+	{
+		$query=array();
+		foreach ($this->fields as $field=>$info)
+		{
+			if($this->is_date($info))
+			{
+				if(empty($this->{$field}))
+				{
+					$query[$field] = NULL;
+				}
+				else
+				{
+					$query[$field] = $this->db->idate($this->{$field});
+				}
+			}
+			else if($this->is_array($info))
+			{
+				$query[$field] = serialize($this->{$field});
+			}
+			else if($this->is_int($info))
+			{
+				$query[$field] = (int) price2num($this->{$field});
+			}
+			else if($this->is_float($info))
+			{
+				$query[$field] = (double) price2num($this->{$field});
+			}
+			elseif($this->is_null($info))
+			{
+				$query[$field] = (is_null($this->{$field}) || (empty($this->{$field}) && $this->{$field}!==0 && $this->{$field}!=='0') ? null : $this->{$field});
+			}
+			else
+			{
+				$query[$field] = $this->{$field};
+			}
+		}
+		
+		return $query;
+	}
 
 	/**
 	 * Create object into database
@@ -4674,28 +4827,82 @@ abstract class CommonObject
 	 */
 	public function createCommon(User $user, $notrigger = false)
 	{
-	    foreach ($this->fields as $k => $v) {
 	    
-	        $keys[] = $k;
-	        $values[] = $this->quote($v);
-	         
+	    $fields = array_merge(array('datec'=>$this->db->idate(dol_now())), $this->set_save_query());
+	    
+	    foreach ($fields as $k => $v) {
+	    	
+	    	$keys[] = $k;
+	    	$values[] = $this->quote($v);
+	    	
 	    }
-	    
-	    $sql = 'INSERT INTO '.MAIN_DB_PREFIX.$table.'
+	    $sql = 'INSERT INTO '.MAIN_DB_PREFIX.$this->table_element.'
 					( '.implode( ",", $keys ).' )
 					VALUES ( '.implode( ",", $values ).' ) ';
-	    
-	    $res = $this->query($sql);
+	    $res = $this->db->query( $sql );
 	    if($res===false) {
-	    
-	        return false;
+	    	
+	    	return false;
 	    }
 	    
 	    // TODO Add triggers
-	    	  
-	    return true;	    
+	    
+	    return true;
+	    
 	}
 	
+	/**
+	 * Function to load data into current object this
+	 *
+	 * @param   stdClass    $obj    Contain data of object from database
+	 */
+	private function set_vars_by_db(&$obj)
+	{
+		foreach ($this->fields as $field => $info)
+		{
+			if($this->is_date($info))
+			{
+				if(empty($obj->{$field}) || $obj->{$field} === '0000-00-00 00:00:00' || $obj->{$field} === '1000-01-01 00:00:00') $this->{$field} = 0;
+				else $this->{$field} = strtotime($obj->{$field});
+			}
+			elseif($this->is_array($info))
+			{
+				$this->{$field} = @unserialize($obj->{$field});
+				// Hack for data not in UTF8
+				if($this->{$field } === FALSE) @unserialize(utf8_decode($obj->{$field}));
+			}
+			elseif($this->is_int($info))
+			{
+				$this->{$field} = (int) $obj->{$field};
+			}
+			elseif($this->is_float($info))
+			{
+				$this->{$field} = (double) $obj->{$field};
+			}
+			elseif($this->is_null($info))
+			{
+				$val = $obj->{$field};
+				// zero is not null
+				$this->{$field} = (is_null($val) || (empty($val) && $val!==0 && $val!=='0') ? null : $val);
+			}
+			else
+			{
+				$this->{$field} = $obj->{$field};
+			}
+			
+		}
+	}
+	
+	/**
+	 * Function to concat keys of fields
+	 *
+	 * @return string
+	 */
+	private function get_field_list()
+	{
+		$keys = array_keys($this->fields);
+		return implode(',', $keys);
+	}
 	
 	/**
 	 * Load object in memory from the database
@@ -4708,6 +4915,32 @@ abstract class CommonObject
 	public function fetchCommon($id, $ref = null)
 	{
 	    
+		if (empty($id) && empty($ref)) return false;
+		
+		$sql = 'SELECT '.$this->get_field_list().', datec, tms';
+		$sql.= ' FROM '.MAIN_DB_PREFIX.$this->table_element;
+		
+		if(!empty($id)) $sql.= ' WHERE rowid = '.$id;
+		else  $sql.= ' WHERE ref = \''.$this->quote($ref).'\'';
+		
+		$res = $this->db->query($sql);
+		if($obj = $this->db->fetch_object($res))
+		{
+			$this->id = $id;
+			$this->set_vars_by_db($obj);
+			
+			$this->datec = $this->db->idate($obj->datec);
+			$this->tms = $this->db->idate($obj->tms);
+			
+			return $this->id;
+		}
+		else
+		{
+			$this->error = $this->db->lasterror();
+			$this->errors[] = $this->error;
+			return -1;
+		}
+		
 	}
 
 	/**
@@ -4720,34 +4953,36 @@ abstract class CommonObject
 	 */
 	public function updateCommon(User $user, $notrigger = false)
 	{
-	    foreach ($this->fields as $k => $v) {
-	    
-	        if (is_array($key)){
-	            $i=array_search($k, $key);
-	            if ( $i !== false) {
-	                $where[] = $key[$i].'=' . $this->quote( $v ) ;
-	                continue;
-	            }
-	        } else {
-	            if ( $k == $key) {
-	                $where[] = $k.'=' .$this->quote( $v ) ;
-	                continue;
-	            }
-	        }
-	    
-	        $tmp[] = $k.'='.$this->quote($v);
-	    }
-	    $sql = 'UPDATE '.MAIN_DB_PREFIX.$table.' SET '.implode( ',', $tmp ).' WHERE ' . implode(' AND ',$where) ;
-	    $res = $this->query( $sql );
-	    
-	    if($res===false) {
-	        //error
-	        return false;
-	    }
-	    
-	    // TODO Add triggers
-	    
-	    return true;	    
+		$fields = $this->set_save_query();
+		
+		foreach ($fields as $k => $v) {
+			
+			if (is_array($key)){
+				$i=array_search($k , $key );
+				if ( $i !== false) {
+					$where[] = $key[$i].'=' . $this->quote( $v ) ;
+					continue;
+				}
+			} else {
+				if ( $k == $key) {
+					$where[] = $k.'=' .$this->quote( $v ) ;
+					continue;
+				}
+			}
+			
+			$tmp[] = $k.'='.$this->quote($v);
+		}
+		$sql = 'UPDATE '.MAIN_DB_PREFIX.$this->table_element.' SET '.implode( ',', $tmp ).' WHERE rowid='.$this->id ;
+		$res = $this->db->query( $sql );
+		
+		if($res===false) {
+			//error
+			return false;
+		}
+		
+		// TODO Add triggers
+		
+		return true;
 	}
 	
 	/**
@@ -4760,7 +4995,16 @@ abstract class CommonObject
 	 */
 	public function deleteCommon(User $user, $notrigger = false)
 	{
-	    
+		$sql = 'DELETE FROM '.MAIN_DB_PREFIX.$this->table_element.' WHERE rowid='.$this->id;
+		
+		$res = $this->db->query( $sql );
+		if($res===false) {
+			return false;
+		}
+		
+		// TODO Add triggers
+		
+		return true;
 	}
 
 	/**
@@ -4769,11 +5013,11 @@ abstract class CommonObject
 	 * @param string|int	$value	value to protect
 	 * @return string|int
 	 */
-	function quote($value) {
+	protected function quote($value) {
 	
 	    if(is_null($value)) return 'NULL';
 	    else if(is_numeric($value)) return $value;
-	    else return "'".$this->escape( $value )."'";
+	    else return "'".$this->db->escape( $value )."'";
 	
 	}
